[better-structure] Separated out actual app implementation from how it will be accessed in dev and prod environments.
